refactor: rewrite with typescript
